finish publish

+++ util/file.go
@@ -28,17 +28,28 @@ func CopyDir(src string, dst string) {
 
 	for _, file := range files {
 		if file.IsDir() == true {
+			CopyDir(filepath.Join(src, file.Name()), filepath.Join(dst, file.Name()))
 			continue
 		}
 		CopyFile(filepath.Join(src, file.Name()), filepath.Join(dst, file.Name()))
 	}
 }
 
-func ClearDir(path string) {
+func ClearDir(path string, except []string) {
 	files, err := ioutil.ReadDir(path)
 	CheckError(err)
 
 	for _, file := range files {
+		isExcept := false
+		for _, name := range except {
+			if file.Name() == name {
+				isExcept = true
+				break
+			}
+		}
+		if isExcept {
+			continue
+		}
 		joinedPath := filepath.Join(path, file.Name())
 		if file.IsDir() == true {
 			os.RemoveAll(joinedPath)
